+++ e2e/constants.ts
@@ -1,0 +1,2 @@
+export const API = 'http://localhost:8080/api';
+export const VALID_VIEW_ID = '12345678-1234-1234-1234-123456789012';
